% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgg.R
\name{rcgg}
\alias{rcgg}
\alias{dcgg}
\alias{pcgg}
\alias{qcgg}
\title{Compound Geometric Gamma}
\usage{
rcgg(n, mu, sigma, pi)

dcgg(x, mu, sigma, pi, log = FALSE)

pcgg(x, mu, sigma, pi, lower.tail = TRUE, log.p = FALSE)

qcgg(p, mu, sigma, pi, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{number of observations.}

\item{mu}{mean of the serial interval distribution.}

\item{sigma}{standard deviation of the serial interval distribution.}

\item{pi}{success probability to sample the secondary cases.
The value must be in (0,1].}

\item{x}{vector of quantiles representing the period of symptom onset times between
a pair of infector-infectee.}

\item{log, log.p}{logical; if \code{TRUE}, the probabilities/densities \eqn{p} are returned
as \eqn{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE}, the probabilities are returned as \eqn{Pr(X \le x)},
otherwise \eqn{Pr(X > x)}.}

\item{p}{vector of probabilities.}
}
\value{
\code{rcgg} generates random variables, \code{dcgg} returns the density,
\code{pcgg} returns the cumulative density, and \code{qcgg} returns
the quantile given probability \code{p} of CGG distribution. \cr\cr
The length of the result is determined by \code{n} for \code{rcgg}, the length of
\code{x} for \code{dcgg} and \code{pcgg}, and the length of \code{p} for \code{qcgg}.
}
\description{
Random generator, density, cumulative density, and quantile functions
of a Compound Geometric
Gamma (CGG) distribution.
}
\details{
The distribution is to model the observed serial interval distribution coming from
non-coprimary transmission (a primary-secondary transmission or indirect primary-secondary
transmission
with some unknown intermediate unsampled individuals in between). \cr\cr
The density function is given as follows \cr
\deqn{f(t) = \sum_m g(t|m) * Pr(M=m), t > 0, m = 0,1,...}
\eqn{g(t|m)} is the Gamma distributed density
function with shape \eqn{(m+1)\alpha} and rate \eqn{\beta}, and \eqn{Pr(M = m)} is the
probability mass function of Geometric distribution. Here, \eqn{\alpha, \beta} are shape
and rate parameters of the true serial interval distribution. \cr\cr
The quantile function is determined by solving the optimization function as follows
\deqn{arg min_t (F(t)-p)^2,}
where \eqn{F(t)} is the cumulative density function of CGG distribution and \eqn{p}
is a given probability.
}
\examples{
x <- rcgg(10, 4.5, 2, .8)
fx <- dcgg(x, 4.5, 2, .7, T)
Fx <- pcgg(x, 4.5, 2, .7)

}
