% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTC.R
\name{createTC.simOutbreak}
\alias{createTC.simOutbreak}
\title{Create a Transmission Cloud}
\usage{
\method{createTC}{simOutbreak}(outbreak, downsampling_prob, cutoff, dna_model = "N", seed = 1)
}
\arguments{
\item{outbreak}{object of class "simOutbreak"}

\item{downsampling_prob}{probability of sampling infected cases}

\item{cutoff}{genomic distance cutoff}

\item{dna_model}{DNA evolution model; see Details}

\item{seed}{seed number}
}
\value{
The function returns the following outputs:\cr
\itemize{
    \item \code{tt} a linelist of the true transmission pairs.
    \item \code{tc} a data frame of all plausible transmission pairs.
}

The \code{tc} data frame consists of these following columns:\cr
\itemize{
    \item \code{inf.source} IDs of the potential infectors
    \item \code{inf.ID} IDs of the infectees
    \item \code{si} serial interval data
    \item \code{type} transmission type: coprimary, non-coprimary, neither
    \item \code{M} number of unsampled intermediaries in the non-coprimary transmission
    \item \code{d} genomic distance
}
}
\description{
Generate a transmission cloud, a collection of all
plaussible transmission pairs.
}
\details{
There are many options to model the DNA evolution;
see \code{ape::dist.dna}.
}
\examples{
p <- .7 # sampling proportion
epsilon <- 15 # genomic distance cutoff
outbreak <- simOutbreak(genint_params = c(4.5,2))
mytc <- createTC(outbreak, p, epsilon)
plot(mytc$tt,
     thin = F,
     x_axis = "inf.times",
     node_color = "group",
     node_value = 1,
     col_pal = c(unsampled = "black", sampled = "gold"),
     arrow_size = .5,
     node_size = 5,
     edge_width = .5,
     node_width = .5,
     height = 800,
     width = 600,
     label = F)
}
