% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Serial Interval Estimation}
\usage{
estimate(
  x,
  init,
  lower,
  upper,
  ci = 0.95,
  ncore = 1,
  prior.pi = NULL,
  prior.w = NULL,
  .showProgress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{vector/list of vectors of the observed symptom onset interval between a pair
of infector-infectee. See "Details".}

\item{init}{initial values of \eqn{\mu, \sigma, \pi, w}, where \eqn{\pi, \sigma > 0},
\eqn{0 < \pi \le 1} and
\eqn{0 \le w \le 1}. See "Details".}

\item{lower, upper}{bounds on the parameters, cannot be \code{NA} or \code{NULL}. By order,
it is \eqn{\mu, \sigma, \pi, w}.}

\item{ci}{confidence interval level for each estimate. By default, it is 0.95.}

\item{ncore}{number of physical CPUs/cores to be used during the computation. If \code{ncore > 1},
the evaluation is done by parallel.}

\item{prior.pi, prior.w}{vector of Beta distributed parameters for \eqn{\pi, w}. See "Details".}

\item{.showProgress}{logical; should a progress bar be shown?}

\item{control}{a list of control parameters. See "Details".}
}
\value{

}
\description{
An optimization method (MLE/MAP) to estimate serial interval distribution.
}
\details{
If the transmission networks are non-unique, \code{x} must be
expressed as a list, in which each element represents the observed serial interval data
per one transmission network. \cr\cr
\code{init} is a numeric vector (cannot be \code{NA} or \code{NULL}) which
(by order) depicts the initial value of \eqn{\mu, \sigma, \pi, w}. \eqn{\mu, \sigma} are
mean and standard deviation of the Gamma distributed serial interval, \eqn{\pi} is the
success probability to sample the secondary cases in non-coprimary transmissions, see
\code{?cgg}, and \eqn{w} is the probability of non-coprimary transmissions in the data. \cr\cr
\code{prior.pi}, \code{prior.w} supply the parameters of Beta distribution. By default,
\code{prior.pi}, \code{prior.w} are non-informative. If not \code{NULL}, MAP (maximum a
posteriori) method is used to estimate the parameters, otherwise MLE (maximum likelihood
estimation) method is used. \cr\cr
\code{control} argument is a list that can supply any of the following components:
\itemize{
   \item \code{tol} convergence tolerance; iteration is terminated when the absolute
   difference in function value between successive iteration is below \code{tol}.
   Default is 1.e-06.
   \item \code{regsimp} a logical variable indicating whether the starting parameter
   configuration is a regular simplex. Default is \code{TRUE}.
   \item \code{maximize} a logical variable indicating whether the objective function
   should be maximized. Default is \code{FALSE}.
} \cr
The observed serial intervals are assumed coming from two transmission types: coprimary and
non-coprimary transmission. The density is a mixture of the two transmission distribution,
which is given as follows
\deqn{f(t) = w*g(t|\mu,\sigma,\pi) + (1-w)*h(t|\mu,\sigma),  t \ge 0,}
where \eqn{g(t), h(t)} are the densities of Compound Geometric Gamma (CGG) distribution
and Folded Gamma Difference (FGD) distribution, respectively. The density above shares the
same parameters with the true serial interval distribution, i.e. the mean \eqn{\mu} and the
standard deviation \eqn{\sigma}.
}
\examples{
data <- simulateData(100, 4, 2, R0 = 1)
ncore <- parallel::detectCores()
est <- estimate(x = data$transnet$onset_diff, init = c(2,2,.5,.5), lower = rep(0,4), upper = c(10,10,1,1), ncore = ncore)
plot(est)
est
}
