% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgd.R
\name{rfgd}
\alias{rfgd}
\alias{dfgd}
\alias{pfgd}
\alias{qfgd}
\title{Folded Gamma Difference}
\usage{
rfgd(n, mu, sigma)

dfgd(x, mu, sigma, log = FALSE)

pfgd(x, mu, sigma, lower.tail = TRUE, log.p = FALSE)

qfgd(p, mu, sigma, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{number of observations.}

\item{mu}{mean parameter of the serial interval distribution.}

\item{sigma}{standard deviation parameter of the serial interval distribution.}

\item{x}{vector of quantiles representing the period of symptom onset times between
a pair of infector-infectee.}

\item{log, log.p}{logical; if \code{TRUE}, the probabilities/densities \code{p} are returned
as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE}, the probabilities are returned as Pr(X <= \code{x}),
otherwise Pr(X > \code{x}).}

\item{p}{vector of probabilities.}
}
\value{
\code{rfgd} generates random variables, \code{dfgd} returns the density,
\code{pfgd} returns the cumulative density,  and \code{qfgd} returns the quantile
given probability \code{p} of FGD distribution. \cr\cr
The length of the result is determined by \code{n} for \code{rfgd}, the length of
\code{x} for \code{dfgd} and \code{pfgd}, and the length of \code{p} for \code{qfgd}.
}
\description{
Random generator, density, cumulative density, and quantile functions of
a Folded Gamma Difference (FGD) distribution.
}
\details{
The distribution is to model the observed serial interval distribution coming from
primary-coprimary transmission, i.e. a type of transmission where two cases are linked,
but both were infected by the unseen third. \cr\cr
The density comes from Folded Gamma Difference (FGD),
which is derived from Gamma Difference Distribution [1]. The function has non-negative
real number support and is expressed as follows \cr
\deqn{f(t) = 2 \int g(s)*g(s-t) ds,} \cr
for \eqn{t \ge 0} and \eqn{t \le s \le \infty}. Here, \eqn{g(t), t > 0} is the density of
the true serial interval following Gamma distribution having mean \code{mu} and standard deviation
\code{sigma}. \cr\cr
The quantile function is determined by solving the optimization function as follows
\deqn{arg min_t (F(t)-p)^2,}
where \eqn{F(t)} is the cumulative density function of FGD distribution
given probability \code{p}.
}
\examples{
x <- rfgd(10, 4.5, 2)
fx <- dfgd(x, 4.5, 2)
Fx <- pfgd(x, 4.5, 2)
q <- qfgd(Fx, 4.5, 2)

}
\references{
\enumerate{
   \item Bernhard Klar (2015) A note on gamma difference distributions,
Journal of Statistical Computation and Simulation, 85:18, 3708-3715,
DOI: 10.1080/00949655.2014.996566
}
}
