% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBeta.R
\name{findBeta}
\alias{findBeta}
\title{Find Beta-distributed Parameters}
\usage{
findBeta(q, p, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{log.p}{logical; if TRUE, \code{p} is returned as log(p).}
}
\value{
Return the Beta-distributed parameters, shape1 and shape2.
}
\description{
Search for Beta-distributed parameters given quantile and probability.
}
\examples{
par <- c(10, 46)
p <- c(.3, .5, .9)
q <- qbeta(p, par[1], par[2])
findBeta(q, p)

}
