% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getci2.siestim.R
\name{getci2.siestim}
\alias{getci2.siestim}
\title{Estimate Confidence Interval Using Likelihood Ratio}
\usage{
\method{getci2}{siestim}(x, level = 0.95)
}
\arguments{
\item{x}{object of class \code{siestim}.}

\item{level}{confidence interval level; default is 0.95.}
}
\value{
confidence interval matrix.
}
\description{
Estimate the \code{(1-a)}-level confidence interval of each parameter's estimate
from class \code{siestim}.
}
