% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logll}
\alias{logll}
\title{The negative log-likelihood functions}
\usage{
logll(params, dt, prior.pi = NULL, prior.w = NULL, ctr = list())
}
\arguments{
\item{prior.w}{}
}
\value{

}
\description{
The negative log-likelihood functions
}
