% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{opt}
\alias{opt}
\title{The optimization function}
\usage{
opt(params0, dt, lower, upper, prior.pi = NULL, prior.w = NULL, ctr = list())
}
\arguments{
\item{prior.w}{}
}
\value{

}
\description{
The optimization function
}
