% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.siestim.R
\name{plot.siestim}
\alias{plot.siestim}
\title{Plot Serial Interval Estimation}
\usage{
\method{plot}{siestim}(x)
}
\arguments{
\item{x}{object of class \code{siestim}}
}
\value{
The function returns the following plots:
\itemize{
   \item \code{Serial Interval Distribution} depicts the distribution of the Gamma-distributed
   serial interval with mean \code{mu} and standard deviation \code{sigma}. If the transmission
   networks are non-unique, the distribution
   of each network will be drawn in gray color, and the final estimated distribution will be drawn in
   coral color.
   \item \code{Mixture Distribution} depicts the densities of the mixture model and
   its mixture components (coprimary and non-coprimary).
   \item \code{Distributions of Estimates} shows figure of the distribution of each
   estimate. If the transmission networks are non-unique, a histogram will be drawn.
}
}
\description{
Draw plots of serial interval distribution, observed serial interval
distribution, and each parameter estimate.
}
