% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simdata.R
\name{plot.simdata}
\alias{plot.simdata}
\title{Plot Transmission Network}
\usage{
\method{plot}{simdata}(x)
}
\arguments{
\item{x}{object of class \code{simdata}.}
}
\value{
A \code{visNetwork} visualization plot of the generated transmission network. \cr\cr
}
\description{
A function to visualize the generated transmission network by \code{simulateData()}.
}
\details{
The node size is scaled proportionally to the amount of direct/indirect infectees
connected to the case, and the edge length is proportional to the period of symptom onset
times between two connected cases.
}
\examples{
data <- simulateData(n = 100, mu = 4, sigma = 2, R0 = .7)
plot(data)
}
