% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simOutbreak.R
\name{simOutbreak}
\alias{simOutbreak}
\title{Generate SIR Outbreak}
\usage{
simOutbreak(
  initial_sus = 200,
  genint_params,
  outbreak_duration = 100,
  R0 = 2,
  transi_rate = 1e-04,
  transv_rate = transi_rate/2,
  seq_length = 10000,
  minimum_case = 10
)
}
\arguments{
\item{initial_sus}{number of initial susceptibles}

\item{genint_params}{a vector of the serial interval distribution's parameters; see Details.}

\item{outbreak_duration}{outbreak duration}

\item{R0}{reproduction number}

\item{transi_rate}{transition rate; see Details}

\item{transv_rate}{transversion rate; see Details}

\item{seq_length}{sequence length to be generated}

\item{minimum_case}{minimum number of infected cases; see Details}
}
\value{
The function returns the following values: \cr
\itemize{
    \item \code{epidata} a data frame of the true transmission pairs.
    \item \code{aligndata} a list of generated sequence data.
    \item \code{dynam} a data frame of the outbreak dynamics.
}

The \code{epidata} data frame consists of these following columns:\cr
\itemize{
    \item \code{inf.ID} case ID
    \item \code{inf.times} time of infection
    \item \code{rec.times} time of recovery
    \item \code{inf.source} infector's ID
    \item \code{nmut} genomic distance (in SNPs)
    \item \code{si} generation/serial interval
}

The \code{dynam} data frame consists of these following columns:\cr
\itemize{
    \item \code{nsus} number of susceptibles at time t
    \item \code{ninf} number of infected cases at time t
    \item \code{nrec} number of recovery cases at time t
}
}
\description{
Generate an SIR outbreak, together with sequence alignment
data.
}
\details{
\code{genit_params} is expressed in terms of the mean and
standard deviation of the serial interval distribution which is
assumed to be Gamma-distributed random variable. \cr

The transition rate is the rate of mutation from purine (A, G) to
purine or pyrimidine (C, T) to pyrimidine. The transversion rate
is the rate of mutation from purine to pyrimidine or vice versa. \cr

If the epidemic size is less than \code{minimum_case}, the function
will generate another outbreak.
}
\examples{
params <- c(4.5,2) # mean = 4.5 days; std.dev = 2 days
myoutbreak <- simOutbreak(genint_params = params)


}
