% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTC.R
\name{createTC.simOutbreak}
\alias{createTC.simOutbreak}
\title{Create a Transmission Cloud}
\usage{
\method{createTC}{simOutbreak}(outbreak, downsampling_prob, cutoff, dna_model = "N", seed = 1)
}
\arguments{
\item{outbreak}{object of class "simOutbreak"}

\item{downsampling_prob}{probability of sampling infected cases}

\item{cutoff}{genomic distance cutoff}

\item{dna_model}{DNA evolution model; see Details}

\item{seed}{seed number}
}
\value{
The function returns the following outputs:\cr
\itemize{
    \item \code{tt} a linelist of the true transmission pairs.
    \item \code{tc} a data frame of all plausible transmission pairs.
}

The column's name of \code{tt} and \code{tc}:\cr
\itemize{
    \item \code{inf.source} IDs of the potential infectors
    \item \code{inf.ID} IDs of the infectees
    \item \code{inf.times} time of infection
    \item \code{rec.times} time of recovery
    \item \code{nmut} genetic distance (in SNPs)
    \item \code{si} serial interval data
    \item \code{type} transmission type: coprimary, non-coprimary, neither
    \item \code{M} number of unsampled intermediaries in the non-coprimary transmission
    \item \code{d} genomic distance
}
}
\description{
Generate a transmission cloud, a collection of all
plaussible transmission pairs.
}
\details{
There are many options to model the DNA evolution;
see \code{ape::dist.dna}.
}
\examples{
p <- .7 # sampling proportion
epsilon <- 15 # genomic distance cutoff
outbreak <- simOutbreak(genint_params = c(4.5,2))
mytc <- createTC(outbreak, p, epsilon)
}
