% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getci.siestim.R, R/getci2.siestim.R
\name{getci2.siestim}
\alias{getci2.siestim}
\title{Estimate Confidence Interval Using Wald Interval}
\usage{
\method{getci2}{siestim}(x, level = 0.95)

\method{getci2}{siestim}(x, level = 0.95)
}
\arguments{
\item{x}{object of class \code{siestim}.}

\item{level}{confidence interval level; default is 0.95.}
}
\value{
confidence interval matrix.

confidence interval matrix.
}
\description{
Estimate the \code{(1-a)}-level confidence interval of each parameter's estimate
from class \code{siestim}.

Estimate the \code{(1-a)}-level confidence interval of each parameter's estimate
from class \code{siestim}.
}
\details{
For an estimate \eqn{\theta}, the interval is estimated as follows \cr
\deqn{\theta - z_{a/2}*se(\theta), \theta - z_{a/2}*se(\theta)}
where \eqn{z_a} is \code{(1-a)}-th quantile of the standard Normal distribution,
and \eqn{se(\theta)} is the estimated standard error of \eqn{\theta}.
}
